UT: Use mock encryptor
